Find closest matching package

Find closest matching package

FindClosestPackage(pkg, maxdist = 2, auto.select = TRUE,
  username.pref = c("cran", "ropensci", "rstudio", "tidyverse", "hadley",
  "yihui", "RcppCore", "eddelbuettel", "ropenscilabs", "hrbrmstr", "thej022214",
  "bomeara"))

Arguments

pkg
A single package

maxdist
The maximum distance that counts as a match

auto.select
If TRUE, make a best guess in case of multiple equally good

username.pref
In case of matches, user names in order of preference.

Value

pkgs that match the constraints

Details

Inspired by githubinstall::gh_suggest() but allows being pickier about match. The username.pref is based on my guesses on priority: "cran" is a mirror for packages that have been on CRAN at some point, but could have been taken off; "ropensci" and "rstudio" produce really useful packages, etc.

Examples

yearn(TreEvo) # A package on github, not CRAN (yet)

#> [1] "TreEvo not installed, now going to try to find it on CRAN or Bioconductor"

#> Bioconductor version 3.5 (BiocInstaller 1.26.0), ?biocLite for help

#> Installing package into ‘/usr/local/lib/R/3.4/site-library’
#> (as ‘lib’ is unspecified)

#> [1] "Installing TreEvo from CRAN failed, now trying to install from github. First trying from CRAN mirror (which includes packages formerly on CRAN). Note this is case-sensitive."

#> Skipping install of 'prereg' from a github remote, the SHA1 (b8c89850) has not changed since last install.
#>   Use `force = TRUE` to force installation

#> [1] "Successfully installed TreEvo from cran mirror on github"