Do yearn on a single package

Do yearn on a single package

yearn(pkg, maxdist = 2, username.pref = c("cran", "ropensci", "rstudio",
  "tidyverse", "hadley", "yihui", "RcppCore", "eddelbuettel", "ropenscilabs",
  "hrbrmstr", "thej022214", "bomeara"))

Arguments

pkg
The package name (bare text, NOT in quotes)

maxdist
The maximum distance that counts as a match

username.pref
In case of matches, user names in order of preference.

Value

NULL if success, a string describing the problem otherwise

Details

See the readme file. Basically, this tries to load an installed package. If that fails, it then looks (in order) on Bioconductor, CRAN, GitHub's CRAN mirror, and other GitHub repositories for the package and installs it if it can find it.

Examples

not_run({
  yearn(TreEvo) # A package on github, not CRAN (yet)
})